created compiled version of library
