+++ CMakeLists.txt
@@ -4,25 +4,32 @@ project(CXXStateTree VERSION 0.4.0 LANGUAGES CXX)
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-add_library(CXXStateTree INTERFACE)
-target_include_directories(CXXStateTree INTERFACE include)
+file (GLOB_RECURSE SRC_FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp")
+
+add_library(CXXStateTree SHARED ${SRC_FILES})
+target_include_directories(CXXStateTree INTERFACE include ${CMAKE_CURRENT_SOURCE_DIR}/include )
 
 add_executable(basic examples/basic.cpp)
-target_link_libraries(basic PRIVATE CXXStateTree)
-
+target_include_directories(basic PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_link_libraries(basic PUBLIC CXXStateTree)
 
 add_executable(nested examples/nested.cpp)
-target_link_libraries(nested PRIVATE CXXStateTree)
+target_include_directories(basic PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_link_libraries(nested PUBLIC CXXStateTree)
 
 add_executable(export_dot_example examples/export_dot.cpp)
-target_include_directories(export_dot_example PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(export_dot_example PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_link_libraries(export_dot_example PUBLIC CXXStateTree)
 add_executable(export_dot_nested_example examples/export_dot_nested.cpp)
-target_include_directories(export_dot_nested_example PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(export_dot_nested_example PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_link_libraries(export_dot_nested_example PUBLIC CXXStateTree)
 add_executable(export_dot_context_example examples/export_dot_context.cpp)
-target_include_directories(export_dot_context_example PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(export_dot_context_example PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_link_libraries(export_dot_context_example PUBLIC CXXStateTree)
 
 add_executable(context_example examples/context_example.cpp)
-target_include_directories(context_example PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(context_example PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_link_libraries(context_example PUBLIC CXXStateTree)
 
 # GoogleTest setup
 include(FetchContent)
