+++ examples/basic.cpp
@@ -1,13 +1,12 @@
 // File: examples/basic.cpp
 #include <iostream>
-#include "CXXStateTree/StateTree.hpp"
-#include "CXXStateTree/Builder.hpp"
+#include "CXXStateTree/StateTree.h"
 
 using namespace CXXStateTree;
 
 int main()
 {
-    auto machine = Builder()
+    auto machine = StateTree::Builder()
                        .initial("Idle")
                        .state("Idle", [](State &s)
                               { s.on("Start", "Running", nullptr, [](const std::any &)
